Fix error in websocket example
std::min fails to deduce the correct type to use now that
QByteArray::size returns qsizetype instead of always returning int.
A cast is needed on 32-bit now though, so let's cast to 64-bit.

+++ examples/remoteobjects/websockets/common/websocketiodevice.cpp
@@ -83,7 +83,7 @@ void WebSocketIoDevice::close()
 
 qint64 WebSocketIoDevice::readData(char *data, qint64 maxlen)
 {
-    auto sz = std::min(int(maxlen), m_buffer.size());
+    auto sz = std::min(maxlen, qint64(m_buffer.size()));
     if (sz <= 0)
         return sz;
     memcpy(data, m_buffer.constData(), size_t(sz));
